backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
pyjwt==2.8.0
werkzeug==3.0.1
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
gevent==23.9.1
gevent-websocket==0.10.1
